requirements.txt
asgiref==3.8.1
Django==4.2.13
gunicorn==22.0.0
lxml==5.2.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
